requirements.txt
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
boto3==1.14.27
botocore==1.17.27
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
docutils==0.15.2
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
redis==3.5.3
requests==2.24.0
rq==1.4.3
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
